backend/requirements.txt
flask
flask-cors
pillow
pytesseract
transformers
easyocr
opencv-python